# Fix Vulkan CPU/GPU synchronization