--- a/platform/ios/vulkan_context_ios.mm
+++ b/platform/ios/vulkan_context_ios.mm
@@ -57,7 +57,8 @@ Error VulkanContextIOS::window_create(DisplayServer::WindowID p_window_id, Displ
 	return _window_create(p_window_id, p_vsync_mode, surface, p_width, p_height);
 }
 
-VulkanContextIOS::VulkanContextIOS() {}
+VulkanContextIOS::VulkanContextIOS() {
+}
 
 VulkanContextIOS::~VulkanContextIOS() {}