… drivers/vulkan/vulkan_context.cpp