The original code had two bugs:

1. It assumed Swapchain count can be used for proper CPU/GPU synchronization (it's not)
2. It used a double buffer scheme (FRAME_LAG = 2) but the vkWaitForFences/vkResetFences calls were in the wrong order and in the wrong place
3. The swapchain bug forced a quad-buffer scheme with roughly 1 or 2 frames of wait; which hid all the issues

This commit gets rid of `RenderingDeviceVulkan::frame_count` & `RenderingDeviceVulkan::frame` which should've never existed and makes everything depend on `VulkanContext::frame_index` &
`VulkanContext::FRAME_LAG`

See https://github.com/godotengine/godot/issues/80550 for details

Fix #80550
Likely fixes https://github.com/godotengine/godot/issues/81670.

## Specific things that need testing:

Aside from overall stability and ensure there are no regressions

- [ ] See if there is an improvement for DEVICE_LOST users. #71929
- [ ]  See if users who reported higher latency/lag vs OpenGL renderer report improvements #71795
- [x] Test if performance changes (for better or worse)
- [x] Test FRAME_LAG = 3. The buggy behavior had the latency disadvantages of triple buffering with the performance disadvantages of double buffering. Perhaps we can get a speed boost with the same latency as before by using FRAME_LAG = 3

**Note:** This PR may uncover [bugs that were previously hidden](https://github.com/godotengine/godot/issues/80570) or that were rare to reproduce. This may give the illusion that the code has become more unstable when in reality it's just uncovering hidden flaws.