@BastiaanOlij 2023-08-13
I need to test this sometime during the week with VR. I've had a lot of timing issues because our frame timing seemed to clash with OpenXRs `xrWaitFrame` logic, or at least thats always been my conclusion, but possibly this is the real issue.