@darksylinc 2023-08-13
OK I have evidence that lag has been reduced.

I couldn't use OBS because it introduced atrocious lag when enabled. I had to use an external camera.

I've ran multiple non-scientific tests and it felt better but I was heavily biased (I *want* to claim my version does something); so I had to 

# Methodology

1. Xubuntu 20.04 XFCE. No Compositor (Window Manager Tweaks -> Compositor -> Enable display compositing must be off)
2. X11 - Vulkan (Forward+) - dedicated AMD Radeon RX 6800 XT - AMD Ryzen 9 5900X 12-Core Processor (24 Threads)
3. Using RADV driver. Mesa 23.0.4 - kisak-mesa PPA
4. TearFree enabled (xrandr --output DisplayPort-1 -set TearFree on)
5. LG FLATRON E2342 1920x1080@60hz
6. Open near-empty project ([PopupCrash](https://github.com/godotengine/godot/files/8459740/PopupCrash.zip)).
7. Camera used for recording: Xiaomi POCO F2 set to 1080p@60 FPS
8. VSync on
9. Update Continuously is off
10. Mouse movement is emulated with the following bash script (use `sudo apt install xautomation` if not installed):
```bash
while true
do
	xte 'mousemove 100 100'
	xte 'mousemove 100 125'
	xte 'mousemove 100 150'
	xte 'mousemove 100 175'
	xte 'mousemove 100 200'
	xte 'mousemove 100 225'
	xte 'mousemove 100 250'
	xte 'mousemove 100 275'
	xte 'mousemove 100 300'
	xte 'mousemove 100 325'
	xte 'mousemove 100 350'
	xte 'mousemove 100 325'
	xte 'mousemove 100 300'
	xte 'mousemove 100 275'
	xte 'mousemove 100 250'
	xte 'mousemove 100 225'
	xte 'mousemove 100 200'
	xte 'mousemove 100 175'
	xte 'mousemove 100 150'
	xte 'mousemove 100 125'
	xte 'mousemove 100 100'
done
```

These are the results (I've had to re-encode the videos because they were too big for Github):

**Before PR**
https://github.com/godotengine/godot/assets/3395130/63a681df-98e6-4836-aa24-cf0e43977b58

**After PR**
https://github.com/godotengine/godot/assets/3395130/d5b40e06-23f5-4742-8392-5cd9cb674bd8


**We can notice:**

1. Before PR the highlight lags behind a lot
2. Before PR the hightlight is almost always 9 highlights behind (when the cursor is in "New Inherited Scene..." the highlight is in "Quick Open Scene...")
3. After PR lags behind a lot less
4. Even though in a few occasions it manages to be 9 highlights behind, it happens twice in the video, while most often is only at most 7 highlights behind; while most of the time is 3 highlights behind.

Overall I believe the data is consistent with latency being reduced by 1 frame (16 ms). I encourage others to do their own research and validate the result.

Unfortunately I only run the test once (I'm a bit of a hurry, tomorrow it's election day).

IMHO it feels noticeable snappier too. Of course it will never reach the level of snappiness you can get with VSync Off, however theoretically the latency with VSync off should be reduced as well.