@darksylinc 2023-08-20
I'm adding new functionality.

In `Window -> V-Sync -> Buffer Count` users can now set the FRAME_LAG (which was previously hardcoded) anywhere in range [1; 4].

Since it must be provided at very early initialization, I didn't add a gdscript interface for it. That can be added & figured out later by anyone else.

![Godot](https://github.com/godotengine/godot/assets/3395130/0d93b417-9104-4f31-bdc0-8ec38f5b1dee)

The plumbing for that is done now (you may have feedback for changes of course, though).

This setting should also address @clayjohn concerns; although they may be misplaced.

Even if Godot has no windows, `VulkanContext::swap_buffers()` still must be called. Whether frame_count is 1 or 2 is irrelevant to whether there are windows.
frame_count just means how much the CPU is allowed to get ahead of the GPU.