@Calinou 2023-08-20
> Since it must be provided at very early initialization, I didn't add a gdscript interface for it. That can be added & figured out later by anyone else.

This may likely require something like https://github.com/godotengine/godot-proposalsI/issues/6423 to be changeable at run-time.