@darksylinc 2023-08-26
OK I've updated the PR with `buffer_count` and `swapchain_count` separately.

The documentation includes a link to the [V-Sync Simulator](https://darksylinc.github.io/vsync_simulator/) I wrote to explain how they work.

I think the PR is ready to merge. If everyone's ok with the review, I'll rebase all the commits to squash them so it can be merged.