@darksylinc 2023-09-16
I had to add more code in light of https://github.com/godotengine/godot/issues/81670 to prevent destroying potentially-uninitialized window' data.

The isolated code changes [can be evaluated here](https://github.com/darksylinc/godot/commit/8ea13496aa1f1017e7ab138b61d7fd9be521fbbf). I wanted to submit them as separate PRs but there were too many merge conflicts otherwise.