@TempoLabGames 2023-09-23
I've built and tested this using https://github.com/TempoLabGames/UnityRenderingLatencyDemo

Summary: This change makes a significant difference to ExclusiveFullscreen, but no difference to Windowed/Fullscreen.

I've included measurements against Godot 4.1.1, your latest commit (my changes to build the .NET editor are available at https://github.com/TempoLabGames/godot/commits/matias-vulkan-sync-fix), and Unity's Vulkan and Direct3D11 implementations for reference.

(Note Godot's ExclusiveFullscreen is equivalent to Unity's FullScreenWindow; **not** Unity's ExclusiveFullScreen.)

Godot 4.1.1 Forward+:
* Windowed: 2-3 frames
* ExclusiveFullscreen: 5 frames

Godot 4302a80 Forward+ (buffer_count=2, swapchain_count=3):
* Windowed: 2-3 frames
* ExclusiveFullscreen: 4 frames

Godot 4302a80 Forward+ (buffer_count=2, swapchain_count=2):
* Windowed: 2-3 frames
* ExclusiveFullscreen: 3 frames

Godot 4302a80 Forward+ (buffer_count=1, swapchain_count=2):
* Windowed: 2-3 frames
* ExclusiveFullscreen: 2 frames (or slightly less)

Unity 2021.3 Vulkan (QualitySettings.maxQueuedFrames=1):
* Windowed: 3 frames
* FullScreenWindow: 5 frames

Unity 2021.3 Direct3D11 (QualitySettings.maxQueuedFrames=1):
* Windowed: 2 frames
* FullScreenWindow: 1 frame