@darksylinc 2023-09-23
I've rebased this branch with latest master and incorporated the latest suggestions.