@TisFoolish 2023-09-23
> (Note Godot's ExclusiveFullscreen is equivalent to Unity's FullScreenWindow; **not** Unity's ExclusiveFullScreen.)

@TempoLabGames Is Unity's ExclusiveFullScreen at 0 frames of delay?

I assume these tests are on Windows (which version?) because of the inclusion of the D3D11 test, unless you're running this through proton? 

What GPU are you using?